simplified stats module
